requirements.txt
gitpython
dirsync
